+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fix bug where in some rare cases a negative INP or LCP value could be reported.
 
+### Other
+
+- Remove `LUX.enablePostBeacon` configuration item.
+
 ## 2024-09-19: v4.0.27
 
 ### Bug fixes